Small refactoring to checkbox
